Move attachments fetching to controller
